Move tests into their own subdirectories
